refactor: remove arrays with keyboard to keyboard.js file, change createKeyboard function
