+++ styles.scss
@@ -21,13 +21,14 @@ body {
     align-items: center;
     box-sizing: border-box;
 
-    &:nth-child(even) {
+    &:nth-child(3n) {
         background: rgba(56, 56, 56, 0.067);
     }
 }
 
 .text {
     margin: 50px auto;
+    height: 20vh;
     width: 80%;
 }
